Sanitise user agent input to matcher
The robotstxt library was designed to be part of a web crawler
and so does not support a user agent input value containing
a version number, such as "curl/8.7.1". Rather than attempt to
implement an unnatural feature in robotstxt, the fix is to
sanitise the user agent before passing it to the matcher.

For simplicity, a private function from robotstxt is copied to
this project, which is ok since both are licensed under Apache v2.0.